+++ assets/javascripts/kitten/components/molecules/cards/summary-card/stories.js
@@ -1,22 +1,99 @@
 import React from 'react'
+import { SummaryCard, useSummaryCardResizeObserver } from './index'
 import {
+  DropdownMenu,
+  EllipsisIcon,
   FlexWrapper,
   Tag,
   Text,
-  DropdownMenu,
-  EllipsisIcon,
-  SummaryCard,
   Title,
-  useSummaryCardResizeObserver,
 } from 'kitten'
+import { DocsPage } from 'storybook/docs-page'
+import { action } from '@storybook/addon-actions'
 
-export const WithContributions = ({ hasImage, hasTitleTag, ...args }) => {
+const argTypes = {
+  show: {
+    name: 'show',
+    description: 'If empty return null',
+    control: { type: 'boolean' },
+  },
+  as: {
+    name: 'as',
+    description: 'HTML element for the Action component',
+    control: { type: 'text' },
+  },
+  actionProps: {
+    name: 'actionProps',
+    description: 'Properties for the Action component',
+    control: { type: 'object' },
+  },
+  hasImage: {
+    name: 'hasImage (story prop)',
+    description: 'Show/hide image, for story purposes.',
+    control: { type: 'boolean' },
+  },
+  hasTitleTag: {
+    name: 'hasTitleTag (story prop)',
+    description: 'Show/hide title tag, for story purposes.',
+    control: { type: 'boolean' },
+  },
+}
+
+const args = {
+  show: true,
+  actionProps: {
+    href: '#',
+    as: 'a',
+    'aria-label': 'A label for the link',
+    onClick: e => {
+      e.preventDefault()
+      action('Clicked')(e)
+    },
+  },
+  hasImage: true,
+  hasTitleTag: true,
+}
+
+export default {
+  component: SummaryCard,
+  title: 'Molecules/Cards/SummaryCard',
+  parameters: {
+    component: SummaryCard,
+  },
+  parameters: {
+    docs: {
+      page: () => (
+        <DocsPage
+          filepath={__filename}
+          filenames={[
+            'index.js',
+            'styles.js',
+            'components/cell.js',
+            'components/image.js',
+            'components/index.js',
+            'components/title-bar.js',
+            'components/title-tag.js',
+          ]}
+          importString="SummaryCard"
+        />
+      ),
+    },
+  },
+  decorators: [story => <div className="story-Container">{story()}</div>],
+  args,
+  argTypes,
+}
+
+export const WithCells = ({ hasImage, hasTitleTag, ...args }) => {
   const { ref, size } = useSummaryCardResizeObserver()
 
   return (
     <div ref={ref}>
       <FlexWrapper gap={50}>
         <div>
+          <Title modifier="septenary" noMargin className="k-u-margin-bottom-double">
+            Owner Contribution
+          </Title>
           <SummaryCard.TitleBar
             values={{
               image: 'Visuel de la contrepartie',
@@ -91,6 +168,9 @@ export const WithContributions = ({ hasImage, hasTitleTag, ...args }) => {
         </div>
 
         <div>
+          <Title modifier="septenary" noMargin className="k-u-margin-bottom-double">
+            Owner Subscription
+          </Title>
           <SummaryCard.TitleBar
             values={{
               image: "Visuel de l'abonnement",
@@ -173,12 +253,89 @@ export const WithContributions = ({ hasImage, hasTitleTag, ...args }) => {
         </div>
 
         <div>
-          <Title
-            modifier="senary"
-            className="k-u-margin-bottom-tripleHalf k-u-margin-top-septuble"
-            noMargin
-          >
-            Abonnements et dons récurrents
+          <Title modifier="septenary" noMargin className="k-u-margin-bottom-double">
+            Contributor Contribution
+          </Title>
+          <SummaryCard.TitleBar
+            values={{
+              image: '',
+              title: 'Projet',
+              description: 'Description',
+              amount: 'Montant',
+              payment: 'Paiement',
+              shipping: 'Livraison',
+              action: '',
+            }}
+            className="k-u-hidden@xs-down k-u-margin-bottom-single"
+            id="SummaryList-contributorContribution"
+            size={size}
+            type="contributorContribution"
+          />
+
+          <SummaryCard {...args} size={size} type="contributorContribution">
+            <SummaryCard.Image className="k-u-hidden@xs-down">
+              <img
+                src={`/kitten-${Math.floor(Math.random() * 10)}.jpg`}
+                alt=""
+              />
+            </SummaryCard.Image>
+
+            <SummaryCard.Cell
+              name="title"
+              aria-describedby="SummaryList-contributorContribution-title"
+            >
+              <Text weight="bold" size="tiny">
+                Contributor Contribution dolor auctor.
+              </Text>
+            </SummaryCard.Cell>
+
+            <SummaryCard.Cell
+              name="description"
+              aria-describedby="SummaryList-contributorContribution-description"
+            >
+              <Text size="tiny">Sollicitudin</Text>
+            </SummaryCard.Cell>
+
+            <SummaryCard.Cell
+              name="amount"
+              aria-describedby="SummaryList-contributorContribution-amount"
+            >
+              <Text weight="regular" size="tiny">
+                10 000 €
+              </Text>
+            </SummaryCard.Cell>
+
+            <SummaryCard.Cell
+              name="payment"
+              aria-describedby="SummaryList-contributorContribution-payment"
+            >
+              <Tag type="success">Validé</Tag>
+            </SummaryCard.Cell>
+
+            <SummaryCard.Cell
+              className="k-u-hidden@xs-down"
+              name="shipping"
+              aria-describedby="SummaryList-contributorContribution-shipping"
+            >
+              <Text weight="regular" size="tiny" className="k-u-block">
+                Nov 2021
+              </Text>
+              <Text size="tiny" className="k-u-block">
+                En main propre
+              </Text>
+            </SummaryCard.Cell>
+
+            <SummaryCard.Cell name="last" className="k-u-align-right">
+              <span className="k-u-size-tiny k-u-link k-u-link-primary1">
+                Détails
+              </span>
+            </SummaryCard.Cell>
+          </SummaryCard>
+        </div>
+
+        <div>
+          <Title modifier="septenary" noMargin className="k-u-margin-bottom-double">
+            Contributor Subscription
           </Title>
           <SummaryCard.TitleBar
             values={{
@@ -197,7 +354,7 @@ export const WithContributions = ({ hasImage, hasTitleTag, ...args }) => {
             type="contributorSubscription"
           />
 
-          <SummaryCard className="k-u-margin-bottom-single" {...args} size={size} type="contributorSubscription">
+          <SummaryCard {...args} size={size} type="contributorSubscription">
             <SummaryCard.Image className="k-u-hidden@xs-down">
               <img
                 src={`/kitten-${Math.floor(Math.random() * 10)}.jpg`}
@@ -267,223 +424,24 @@ export const WithContributions = ({ hasImage, hasTitleTag, ...args }) => {
               </span>
             </SummaryCard.Cell>
           </SummaryCard>
-          <SummaryCard {...args} size={size} type="contributorSubscription">
-            <SummaryCard.Image className="k-u-hidden@xs-down">
-              <img
-                src={`/kitten-${Math.floor(Math.random() * 10)}.jpg`}
-                alt=""
-              />
-            </SummaryCard.Image>
-
-            <SummaryCard.Cell
-              name="title"
-              aria-describedby="SummaryList-contributorSubscription-title"
-            >
-              <Text weight="bold" size="tiny">
-                Recto Verso : le point de départ de l'aventure en France
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              name="description"
-              aria-describedby="SummaryList-contributorSubscription-description"
-            >
-              <Text size="tiny">Don libre</Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              name="amount"
-              aria-describedby="SummaryList-contributorSubscription-amount"
-            >
-              <Text size="tiny">
-                <strong className="k-u-weight-regular">2 €</strong> / mois
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              className="k-u-hidden@xs-down"
-              name="payment"
-              aria-describedby="SummaryList-contributorSubscription-payment"
-            >
-              <Text size="tiny" className="k-u-block">
-                Jan 2025
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              name="status"
-              aria-describedby="SummaryList-contributorSubscription-status"
-            >
-              <Tag type="success">Actif</Tag>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              className="k-u-hidden@xs-down"
-              name="shipping"
-              aria-describedby="SummaryList-contributorSubscription-shipping"
-            >
-              <Text weight="regular" size="tiny" className="k-u-block">
-                Nov 2021
-              </Text>
-              <Text size="tiny" className="k-u-block">
-                Pas de livraison
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell name="last" className="k-u-align-right">
-              <span className="k-u-size-tiny k-u-link k-u-link-primary1">
-                Gérer
-              </span>
-            </SummaryCard.Cell>
-          </SummaryCard>
         </div>
-
-        <div>
-          <Title
-            modifier="senary"
-            className="k-u-margin-bottom-tripleHalf k-u-margin-top-septuble"
-            noMargin
-          >
-            Contributions à des campagnes
-          </Title>
-          <SummaryCard.TitleBar
-            values={{
-              image: '',
-              title: 'Projet',
-              description: 'Description',
-              amount: 'Montant',
-              payment: 'Paiement',
-              shipping: 'Livraison',
-              action: '',
-            }}
-            className="k-u-hidden@xs-down k-u-margin-bottom-single"
-            id="SummaryList-contributorContribution"
-            size={size}
-            type="contributorContribution"
-          />
-
-          <SummaryCard className="k-u-margin-bottom-single" {...args} size={size} type="contributorContribution">
-            <SummaryCard.Image className="k-u-hidden@xs-down">
-              <img
-                src={`/kitten-${Math.floor(Math.random() * 10)}.jpg`}
-                alt=""
-              />
-            </SummaryCard.Image>
-
-            <SummaryCard.Cell
-              name="title"
-              aria-describedby="SummaryList-contributorContribution-title"
-            >
-              <Text weight="bold" size="tiny">
-                Contributor Contribution dolor auctor.
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              name="description"
-              aria-describedby="SummaryList-contributorContribution-description"
-            >
-              <Text size="tiny">Sollicitudin</Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              name="amount"
-              aria-describedby="SummaryList-contributorContribution-amount"
-            >
-              <Text weight="regular" size="tiny">
-                10 000 €
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              name="payment"
-              aria-describedby="SummaryList-contributorContribution-payment"
-            >
-              <Tag type="success">Validé</Tag>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              className="k-u-hidden@xs-down"
-              name="shipping"
-              aria-describedby="SummaryList-contributorContribution-shipping"
-            >
-              <Text weight="regular" size="tiny" className="k-u-block">
-                Nov 2021
-              </Text>
-              <Text size="tiny" className="k-u-block">
-                En main propre
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell name="last" className="k-u-align-right">
-              <span className="k-u-size-tiny k-u-link k-u-link-primary1">
-                Détails
-              </span>
-            </SummaryCard.Cell>
-          </SummaryCard>
-
-          <SummaryCard {...args} size={size} type="contributorContribution">
-            <SummaryCard.Image className="k-u-hidden@xs-down">
-              <img
-                src={`/kitten-${Math.floor(Math.random() * 10)}.jpg`}
-                alt=""
-              />
-            </SummaryCard.Image>
-
-            <SummaryCard.Cell
-              name="title"
-              aria-describedby="SummaryList-contributorContribution-title"
-            >
-              <Text weight="bold" size="tiny">
-              Dé-mots : imaginer, écrire, rire
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              name="description"
-              aria-describedby="SummaryList-contributorContribution-description"
-            >
-              <Text size="tiny">Le coffret Dé-mots</Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              name="amount"
-              aria-describedby="SummaryList-contributorContribution-amount"
-            >
-              <Text weight="regular" size="tiny">
-                16 €
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              name="payment"
-              aria-describedby="SummaryList-contributorContribution-payment"
-            >
-              <Tag type="success">Validé</Tag>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell
-              className="k-u-hidden@xs-down"
-              name="shipping"
-              aria-describedby="SummaryList-contributorContribution-shipping"
-            >
-              <Text weight="regular" size="tiny" className="k-u-block">
-                Nov 2021
-              </Text>
-              <Text size="tiny" className="k-u-block">
-                En main propre
-              </Text>
-            </SummaryCard.Cell>
-
-            <SummaryCard.Cell name="last" className="k-u-align-right">
-              <span className="k-u-size-tiny k-u-link k-u-link-primary1">
-                Détails
-              </span>
-            </SummaryCard.Cell>
-          </SummaryCard>
-        </div>
-
       </FlexWrapper>
     </div>
   )
 }
+
+export const TitleBar = args => (
+  <div {...args}>
+    <SummaryCard.TitleBar
+      values={{
+        image: 'Visuel de la contrepartie',
+        title: 'Titre de la contrepartie',
+        amount: 'Montant',
+        contributions: 'Contributions',
+        availability: 'Disponibilités',
+      }}
+      className="k-u-hidden@xs-down"
+      id="SummaryList"
+    />
+  </div>
+)